Work on libc_fuzzy
- corrected issue of input mask setting missing (regression)
- now reasoning works again...

+++ c_fuzzy/include/libc_fuzzy/FuzzyKnowledgeBase.h
@@ -27,7 +27,7 @@ public:
 					knowledgeBase)
 	{
 	}
-	unsigned long size();
+	size_t size();
 	void addRule(Node* fuzzyRule);
 	std::map<std::string, BitData>& getVariableMasks();
 	~FuzzyKnowledgeBase();
